Rename and update OSX/Windows removable drive plugins
